Aula 3 - Criando Usuário e Autenticações

+++ app/views/task_views.py
@@ -1,7 +1,9 @@
+from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse
 from django.shortcuts import render, redirect
-from .forms import TaskForm
-from .entities.task import Task
-from .services import task_service
+from ..forms import TaskForm
+from app.entities.task import Task
+from ..services import task_service
 
 # Create your views here.
 
@@ -14,9 +16,16 @@ Template -> camada de View
 '''
 
 
+# com o @login_required é um decorator que verifica se usuário esta logado
+# se não estiver ele é redirecionado
+# perceba que a página padrão é o account/login; para ajustar isso deve ir em
+# settings.py e criar LOGIN_URL
+
+@login_required()
 def task_list(request):
-    tasks = task_service.task_list()
+    tasks = task_service.task_list(request.user)
     # task_name = 'Assistindo a Treinaweb'
+    # request.user é o usuário logado
 
     return render(request, 'tasks/task_list.html',
                   {"tasks": tasks})
@@ -25,6 +34,7 @@ def task_list(request):
     # o contexto para o template (setAttribute() no java)
 
 
+@login_required()
 def register_task(request):
     # TODO: entender porque o erro não aparece no template
 
@@ -38,7 +48,7 @@ def register_task(request):
             expiration_date = form_task.cleaned_data['expiration_date']
             priority = form_task.cleaned_data['priority']
             new_task = Task(title, description,
-                            expiration_date, priority)
+                            expiration_date, priority, request.user)
 
             task_service.register_task(new_task)
             return redirect('task_list_route')
@@ -48,10 +58,17 @@ def register_task(request):
     return render(request, 'tasks/form_task.html', {"form_task": form_task})
 
 
+@login_required()
 def edit_task(request, id):
     task_db = task_service.task_list_id(id)
+
+    if task_db.user != request.user:
+        return HttpResponse('Não Permitido!')
+        # TODO: melhorar isso, mostar uma mensagem mais arrumada e redirecionar
+
     form_task = TaskForm(request.POST or None, instance=task_db)
     # é como se recebesse o form preenchido do bd.
+
     if form_task.is_valid():
         # se possui os 4 campos válidos
         title = form_task.cleaned_data['title']
@@ -60,7 +77,7 @@ def edit_task(request, id):
         priority = form_task.cleaned_data['priority']
 
         new_task = Task(title, description,
-                        expiration_date, priority)
+                        expiration_date, priority, request.user)
 
         task_service.edit_task(task_db, new_task)
         return redirect('task_list_route')
@@ -68,8 +85,13 @@ def edit_task(request, id):
     return render(request, 'tasks/form_task.html', {"form_task": form_task})
 
 
+@login_required()
 def remove_task(request, id):
     task_db = task_service.task_list_id(id)
+
+    if task_db.user != request.user:
+        return HttpResponse('Não Permitido!')
+
     if request.method == 'POST':
         task_service.remove_task(task_db)
         return redirect('task_list_route')
